WIP : zoom working except when ...
deplaneview area is bigger than browser viewport (e.g. for DE102 with a
wide browser window, the one must scroll to get the bottom of the DE).
In such cases the zoom-in is correct but the zoom-out shift also the DE
upwards ? Most probably solvable with some CSS, but it has nothing to do
with the SVGView zoom and panning handling anyway.